Lint and make test
